agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "langchain-deepseek>=0.1.4",
    "langchain[openai]>=0.3.27",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
